module splitting to shorten main
